Compatibility: action_kill_object
When a Game Maker: Studio project is imported into Game Maker Studio 2, several changes are made to ensure maximum compatibility with the new system. Among those changes are additional scripts added to replace functionality that was either lost or slightly modified in the transition. This is an attempt to purify the editable of these conversions by employing the native GML code directly.

All references to action_kill_object have been replaced with instance_destroy. This is because action_kill_object simply re-directs to instance_destroy, making it redundant.

(The line endings in the compatibility report were also corrected.)