Compatibility: action_inherited
When a Game Maker: Studio project is imported into Game Maker Studio 2, several changes are made to ensure maximum compatibility with the new system. Among those changes are additional scripts added to replace functionality that was either lost or slightly modified in the transition. This is an attempt to purify the editable of these conversions by employing the native GML code directly.

All references to action_inherited have been replaced with event_inherited. This is because action_inherited simply re-directs to event_inherited, making it redundant.