Compatibility: view functionality
When a Game Maker: Studio project is imported into Game Maker Studio 2, several changes are made to ensure maximum compatibility with the new system. Among those changes are additional scripts added to replace functionality that was either lost or slightly modified in the transition. This is an attempt to purify the editable of these conversions by employing the native GML code directly.

While the concept of views still exists in Game Maker Studio 2, most of the view functionality is delegated to the new camera functions, and the older view functions have been removed. The resulting compability scripts are complex lookups that cover all of the imported view code. These scripts have been removed and a preliminary sweep was performed to incorporate the camera functionality in-place. An additional sweep will be executed later to optimize the changes and take full advantage of the new functionality.

(Indirectly, this also fixes a typo that was found in the base code that could potentially cause actions from other players to freeze or otherwise affect the main view during play.)