Compatibility: depth and layering
When a Game Maker: Studio project is imported into Game Maker Studio 2, several changes are made to ensure maximum compatibility with the new system. Among those changes are additional scripts added to replace functionality that was either lost or slightly modified in the transition. This is an attempt to purify the editable of these conversions by employing the native GML code directly.

In Game Maker Studio 2, room instances take on the depth of the layer they are assigned to in a room rather than that of their corresponding object, which no longer have a depth assigned. Upon importing from Game Maker: Studio, GMS2 creates a lookup table of each object's original depth to be referenced when calling instance_create. It also adds layers to each pre-established room and assigns room instances based on their corresponding object's original depth. With this update, a more streamlined system of layers has been established, and instances are either aligned with these layers or instantiated at a depth relative to other instances.