Compatibility: __init_action
When a Game Maker: Studio project is imported into Game Maker Studio 2, several changes are made to ensure maximum compatibility with the new system. Among those changes are additional scripts added to replace functionality that was either lost or slightly modified in the transition. This is an attempt to purify the editable of these conversions by employing the native GML code directly.

__init_action doesn't appear to be necessary and is the last of the compatibility scripts to be removed. With this update, the template now uses all native Game Maker Studio 2 code.