Minor updates (7)
When using the "draw_line" functions, GMS2 automatically sets up a vertex format that excludes texture coordinates. Because nearly all the shaders used by the template require texture coordinates as input, this resulted in a mismatch between the vertex format and the shader format. The result is that frequently, the following error messages would appear in the developer console while playing a level: "Could not generate input layout (is there a mismatch between your shader and vertex format?)", followed by "Draw Failed due to invalid input layout" repeated ad nauseum.

The main culprit of this issue was the draw code employed to display a representation of the player's bounding boxes. This code has been reworked such that there is no longer a format mismatch, elminiating the errors generated.