Bonus stage update (10)
The bonus portal spawn behavior was originally adapted for multiple players, but in retrospect, this decision was not correct. The bonus stage should not be accessible in split-screen multiplayer, and in single player, an AI-controlled player should not be able to prematurely activate a checkpoint,
 using their own ring count to determine if the portal should spawn.
Both issues have been addressed, and now neither scenario is possible. The included object CompetitionLampPost and script player_react_to_competition_lamp_post are stubs to be expanded upon later when a fully fleshed out competition mode is implemented.
(Also updated GlowingSpheresLargeSphere so only player1 plays SphereOrbitSound.)
Thanks to @Bluecoreg for the suggestions.