Refactor and improve touch navigation
This commit is focused on making navigating the File Select menu easier
and re-evaluating which navigation elements are relevant only to it.

Horizontal and vertical swipes can be successfully performed much
slower using a dragging motion, whereas previously it required a fast
flick gesture that felt less responsive. This change only applies to
the File Select menu; the swipe controls previously introduced to
GameMenuPage have been removed entirely as they are not universally
relevant.

To replace the re-purposed swipe controls, a new touch method of
navigation is being introduced to GameMenuChoice. When a menu choice
is not currently selected, tapping that choice (or clicking it with a
mouse, on desktop applications) will directly select it. On the File
Select Menu, this will also move the cursor towards it. Tapping on a
choice already selected will still confirm that selection, as before.

FileSelectMenuCursor now adjusts its own size in response to the menu
selection is it moving towards, rather than it being changed
indirectly by FileSelectMenuPage. Additionally, the cursor height is
now uniform - no more subtle adjustments based on the menu choice.