Fix file creation and touch to skip Splash Screen
Starting a new save file now immediately writes the file to disk.
Previously, this was delegated to the Game End event, where the game
automatically saves all progress upon closing the game (the game also
saves when progressing to the next level.) While this is acceptable on
desktop consoles using the ESC or close buttons, it does not account
for scenarios where closing the application bypasses the Game End
event. This can include hardware or operating system failure, but the
particular event that inspired this commit was closing the application
after sending it to the background on mobile devices.

The Splash Screen was updated to allow skipping by clicking the screen
on desktop application in addition to mobile applications. This
adjustment came about after testing the functionality using the
Virtual Machine, which triggered an error due to an incorrect
variable reference (display_time, instead of duration.) Mysteriously,
this issue was not present when testing using the YoYo Compiler.