Add customizeable touch controls to levels
A touch UI based on the mobile Sonic the Hedgehog ports now appears in
levels. The opacity and scale of the UI can be modified through
GameTouchConfiguration. By default, the touch UI is only enabled on
mobile platforms.

Credit to PorteWeeker for the assets, and a special thanks to Carlos
Ushiromiya for coding assistance.