Enable loop points for queued audio playback
This is part of a series of changes to remove the use of FMODGMS, so
that developers using this framework are not handcuffed to FMOD's
licensing requirements.

Loop point information is now stored with GMSAudioSystem and retrieved
whenever music is queued. This, and some adjustments to GMSGameMusic,
allows queued audio to loop from any point other than the beginning.
GameMaker Studio 2 is unable to return a consistent track position
when the audio is streamed, however, so it is recommended to set any
music using the loop points as uncompressed and not streamed.

To take advantage of this, BonusStageMusic.ogg was removed from the
included files and instead added as a native GameMaker Studio 2 audio
resource to be uncompressed on load.