Switch Studiopolis music from FMODGMS to native
Studiopolis1Music.ogg was removed from the included files and instead
added as a native GameMaker Studio 2 audio resource to be uncompressed
on load. This is part of a series of changes to remove the use of
FMODGMS, so that developers using this framework are not handcuffed to
FMOD's licensing requirements.