Clarify FMODGMS licensing
FMODGMS uses FMOD, which has its own licensing restrictions. Although the newest versions of this software no longer employ FMODGMS, it is important that users are fully aware of the licensing terms, should they choose to use any version that still contains it. A small section has been added to the readme that should help point users in the right direction.