Update credits list
Template assets (art, sound and code) have been derived from many
different sources, some of which have not been properly credited
either due to negligence or because the source did not require it. The
credits list has been updated to include all known contributors, and
re-organized to better reflect their roles. To this end, the credits
font was also updated to properly render the '-' character.