Fix Tails getting stuck at upper camera boundary
The playable character Tails' flight ability works by applying a
negative force of gravity to lift the character into the air. When
flying against a ceiling, this force is changed to a positive number so
the character will fall back to the ground and not get stuck. This
behavior now also applies to the upper boundary of the camera.

Fixes #9