Fix ambiguous use of `self` keyword
Prior to GameMaker Studio 2 version 2.3, using the keyword `self` on its
own was treated as equivalent to `self.id` by the compiler. As of 2.3,
this now returns an instance struct, which broke any functionality that
expected `self.id`. That functionality has been changed to reference
`self.id` directly instead of `self` on its own.