Fix parallax vertical scrolling in YYC
The various level backgrounds were no longer scrolling vertically when
compiling the template to YoYo Compiler (YYC) output. The issue has been
fixed, although it is unclear what was causing the issue as the related
code is unchanged from 2.2.5. (It appears to have something to do with
differences in resolution of the 'div' operator.)

Some minor optimization and reformatting was also performed.

Fixes: #12