Move all scripts to "Deprecated" folder
GameMaker Studio 2 version 2.3 introduced functions and converted all
script resources into scripts containing functions. These "Script
Functions" are not optmized for space or execution. This is an
opportunity to revisit all the script functionality and improve upon it.
Script functionality will be moved out of the deprecated folder and into
new, dedicated scripts as they are revised.