Condense all shader functions
The function `color_get_array_rgb` takes a 32-bit color value and splits
it into an array of color channels (red, green and blue.) This was used
to refactor the function `shader_set_uniform_color_array`, and both
were moved to the "ShaderFunctions" script.

The function `shader_set_uniform_color` was unused, so it was removed
from the project entirely.