Expand range testing functionality
The function `includes` was created as a counter-part to the existing
function `in_range`; it tests whether a given value is part of a bounded
range inclusively, whereas `in_range` tests against range exclusively.
To further clarify this distinction, the latter function was renamed
`between`.

Some slight adjustments were also made to the JavaDocs of these
functions.