Apply range inclusivity tests more liberally
Almost all range tests in the template have thus far been exclusive
tests (ie. using the function `between`) of the range {-1, x}. In these
cases, testing a value between -1 and 0 can result in a false
positive. To avoid this and better express the intent of these tests,
they were replaced with inclusive tests (ie. using the function
`includes`) of the range {0, x - 1}. (The respective code was also
reformated.)